consolidate types and util implementations
