--- src/types/Namespace.ts
@@ -1,12 +1,0 @@
-export enum Namespace {
-  COMPONENTS = "components",
-  INTEGRATIONS = "integrations",
-  MARKETPLACE = "integration-marketplace",
-  MARKETPLACE_INTEGRATION_ALERT_MONITORS = "integrations.id.alert-monitors",
-  MARKETPLACE_INTEGRATION_CONFIGURATION = "integrations.id.configurations",
-  MARKETPLACE_INTEGRATION_EXECUTIONS = "integrations.id.executions",
-  MARKETPLACE_INTEGRATION_LOGS = "integrations.id.logs",
-  MARKETPLACE_INTEGRATION_SUMMARY = "integrations.id",
-  MARKETPLACE_INTEGRATION_TEST = "integrations.id.test",
-  MARKETPLACE_NOT_FOUND = "app.marketplace-not-found",
-}
