--- src/types/NamespacedSharedAndSharedAndUniquePhrases.ts
@@ -1,5 +1,0 @@
-import { SharedAndUniquePhrases } from "./SharedAndUniquePhrases";
-import { NamespacedSharedPhrases } from "./NamespacedSharedPhrases";
-
-export type NamespacedSharedAndSharedAndUniquePhrases =
-  Partial<NamespacedSharedPhrases> & SharedAndUniquePhrases;
