--- src/types/NamespacedSharedAndUniquePhrases.ts
@@ -1,6 +1,0 @@
-import { NamespacedSharedPhrases } from "./NamespacedSharedPhrases";
-import { UniquePhrases } from "../lib/unique";
-
-export type NamespacedSharedAndUniquePhrases<
-  T extends unknown = UniquePhrases
-> = Partial<NamespacedSharedPhrases> & T;
