--- src/types/NamespacedSharedPhrases.ts
@@ -1,6 +1,0 @@
-import { Namespace } from "./Namespace";
-import { SharedPhrases } from "../lib/shared";
-
-export type NamespacedSharedPhrases<T extends unknown = SharedPhrases> = {
-  [K in keyof T as `${Namespace}__${string & K}`]: T[K];
-};
