--- src/types/Phrases.ts
@@ -1,9 +1,0 @@
-import { NamespacedSharedAndSharedAndUniquePhrases } from "./NamespacedSharedAndSharedAndUniquePhrases";
-import { SharedAndUniquePhrases } from "./SharedAndUniquePhrases";
-
-export type Phrases = Partial<NamespacedSharedAndSharedAndUniquePhrases>;
-
-export type Phrase = keyof SharedAndUniquePhrases;
-
-export type PhraseWithNamespace =
-  keyof NamespacedSharedAndSharedAndUniquePhrases;
