--- src/types/PickByType.ts
@@ -1,3 +1,0 @@
-export type PickByType<T, V> = {
-  [P in keyof T as T[P] extends V | undefined ? P : never]: T[P];
-};
