--- src/types/SharedAndUniquePhrases.ts
@@ -1,4 +1,0 @@
-import { SharedPhrases } from "../lib/shared";
-import { UniquePhrases } from "../lib/unique";
-
-export type SharedAndUniquePhrases = SharedPhrases & UniquePhrases;
