--- src/types/SimplePhrase.ts
@@ -1,6 +1,0 @@
-import { Phrases } from "./Phrases";
-import { PickByType } from "./PickByType";
-
-export type SimplePhrase = string;
-
-export type SimplePhrases = PickByType<Phrases, SimplePhrase>;
