--- src/types/index.ts
@@ -1,11 +1,0 @@
-export { ComplexPhrase, ComplexPhrases } from "./ComplexPhrase";
-export { Namespace } from "./Namespace";
-export { NamespacedSharedAndSharedAndUniquePhrases } from "./NamespacedSharedAndSharedAndUniquePhrases";
-export { NamespacedSharedAndUniquePhrases } from "./NamespacedSharedAndUniquePhrases";
-export { NamespacedSharedPhrases } from "./NamespacedSharedPhrases";
-export { Phrases, Phrase, PhraseWithNamespace } from "./Phrases";
-export { PickByType } from "./PickByType";
-export { SharedAndUniquePhrases } from "./SharedAndUniquePhrases";
-export { SharedPhrases } from "../lib/shared";
-export { SimplePhrase, SimplePhrases } from "./SimplePhrase";
-export { UniquePhrases } from "../lib/unique";
