--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export { getComplexPhase } from "./getComplexPhase";
-export { isComplexPhrase } from "./isComplexPhrase";
-export { isSimplePhrase } from "./isSimplePhrase";
