--- src/utils/isComplexPhrase.ts
@@ -1,6 +1,0 @@
-import { ComplexPhrase } from "../types/ComplexPhrase";
-import { SimplePhrase } from "../types/SimplePhrase";
-
-export const isComplexPhrase = (
-  phrase: SimplePhrase | ComplexPhrase
-): phrase is ComplexPhrase => typeof phrase === "object";
